Figured it out

+++ test_rng.cpp
@@ -21,23 +21,15 @@ inline void print_vec(float *a)
 inline bool test_xorshift(int niter=10000)
 {
     random_device rd;
-    // float rn1 = rd();
-    // float rn2 = rd();
-    // float rn3 = rd();
-    // float rn4 = rd();
-    // float rn5 = rd();
-    // float rn6 = rd();
-    // float rn7 = rd();
-    // float rn8 = rd();
-    float rn1 = fast_rng::rd64(rd);
-    float rn2 = fast_rng::rd64(rd);
-    float rn3 = fast_rng::rd64(rd);
-    float rn4 = fast_rng::rd64(rd);
-    float rn5 = fast_rng::rd64(rd);
-    float rn6 = fast_rng::rd64(rd);
-    float rn7 = fast_rng::rd64(rd);
-    float rn8 = fast_rng::rd64(rd);
-
+    uint64_t rn1 = fast_rng::rd64(rd);
+    uint64_t rn2 = fast_rng::rd64(rd);
+    uint64_t rn3 = fast_rng::rd64(rd);
+    uint64_t rn4 = fast_rng::rd64(rd);
+    uint64_t rn5 = fast_rng::rd64(rd);
+    uint64_t rn6 = fast_rng::rd64(rd);
+    uint64_t rn7 = fast_rng::rd64(rd);
+    uint64_t rn8 = fast_rng::rd64(rd);
+          
     fast_rng::vec_xorshift_plus a(_mm256_setr_epi64x(rn1, rn3, rn5, rn7), _mm256_setr_epi64x(rn2, rn4, rn6, rn8));
     float vrn_vec[8];
     rng_helpers::xorshift_plus b(rn1, rn2, rn3, rn4, rn5, rn6, rn7, rn8);
